% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthors_analysis.R
\name{build_author_network}
\alias{build_author_network}
\title{A function used to build collaborative relationships between authors.}
\usage{
build_author_network(author_info)
}
\arguments{
\item{author_info}{author information.}
}
\value{
A network.
}
\description{
A function used to build collaborative relationships between authors.
}
\examples{
\donttest{
myauthor <- NULL
myauthor <- strsplit(author, split = ',')
for (i in length(myauthor)) {
  myauthor[[i]][1] <- paste0(' ', myauthor[[i]][1])
}
g <- build_author_network(myauthor)
}
}
