% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cited_analysis.R
\name{cited_extract}
\alias{cited_extract}
\title{Cited Analysis.}
\usage{
cited_extract(
  root =
    "https://pubmed.ncbi.nlm.nih.gov/?term=(Wang,\%20Xin[Author])\%20AND\%20(Zhang,\%20Yan[Author])&size=20&page=",
  page_number = 1,
  paper_numbers = 1
)
}
\arguments{
\item{root}{common web link part.}

\item{page_number}{page number.}

\item{paper_numbers}{paper number.}
}
\value{
The paper cited details
}
\description{
Cited Analysis.
}
\examples{
\donttest{
cited_doi1 <- list()
cited_doi0 <- list()
mytest <- cited_extract(page_number = 1, paper_numbers = 20)
}
}
