% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_collection.R
\name{crossref_citation}
\alias{crossref_citation}
\title{Obtain information such as citation frequency of literature through CrossRef.}
\usage{
crossref_citation(mydoi)
}
\arguments{
\item{mydoi}{paper DOI.}
}
\value{
Number of citations, etc.
}
\description{
Obtain information such as citation frequency of literature through CrossRef.
}
\examples{
\donttest{
mydoi <- c("10.1016/j.metabol.2020.154171", "10.3390/metabo10020048", "12.3390/metabo10020048")
crossref_citation(mydoi = mydoi)
crossref_citation_table
}
}
