% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{fanyi}
\alias{fanyi}
\title{Literature translation function based on Baidu Translation API.}
\usage{
fanyi(
  q = "偉大な中国人民万歳",
  from = "auto",
  to = "zh",
  appid = "",
  key = "",
  encoding = "utf-8"
)
}
\arguments{
\item{q}{Text to be translated.}

\item{from}{The original language of the text.}

\item{to}{Target language.}

\item{appid}{Your Baidu Translation API ID.}

\item{key}{Your Baidu Translation API key.}

\item{encoding}{Encoding method.}
}
\value{
Translation results.
}
\description{
Literature translation function based on Baidu Translation API.
}
\examples{
\donttest{
fanyi(appid = "xxxxxxxxxxxxxx", key = "xxxxxxxxxxxxxx")
}
}
