% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_download.R
\name{get_scihub_url}
\alias{get_scihub_url}
\title{Obtain the download link for sci hub based on DOI.}
\usage{
get_scihub_url(doi)
}
\arguments{
\item{doi}{The paper doi.}
}
\value{
download link.
}
\description{
Obtain the download link for sci hub based on DOI.
}
\examples{
\donttest{
get_scihub_url(doi = "10.1038/nrmicro3564")
}
}
