% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extract.R
\name{github_link_extract}
\alias{github_link_extract}
\title{Detect and extract github links for a single page.}
\usage{
github_link_extract(
  pdf_text = NULL,
  github_word = NULL,
  git_page_num = NULL,
  filebase = NULL
)
}
\arguments{
\item{pdf_text}{PDF file name.}

\item{github_word}{github related word.}

\item{git_page_num}{numeric.}

\item{filebase}{file base name.}
}
\value{
github link.
}
\description{
Detect and extract github links for a single page.
}
\examples{
\donttest{
github_link_extract()
}
}
