% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscholar_analysis.R
\name{gscholar_analysis}
\alias{gscholar_analysis}
\title{Obtain citation information from Google Scholar.}
\usage{
gscholar_analysis(paper_title = NULL)
}
\arguments{
\item{paper_title}{The paper title.}
}
\value{
Citation information.
}
\description{
Obtain citation information from Google Scholar.
}
\examples{
\donttest{
# Enable proxy
r.proxy::proxy()
# disable proxy
r.proxy::noproxy()
paper_title <- "The hematologic consequences of obesity"
gscholar_analysis(paper_title = paper_title)
cited_title
}
}
