% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{keyword_plot}
\alias{keyword_plot}
\title{Keyword frequency statistics.}
\usage{
keyword_plot(words, size = 0.7, shape = "circle")
}
\arguments{
\item{words}{keyword.}

\item{size}{size.}

\item{shape}{shape.}
}
\value{
Word cloud diagram
}
\description{
Keyword frequency statistics.
}
\examples{
\dottest{
keyword_plot(final_table$Keyword)
}
}
