% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extract.R
\name{pdf_doi_extract}
\alias{pdf_doi_extract}
\title{Extract DOI from PDF.}
\usage{
pdf_doi_extract(pdf_text = NULL)
}
\arguments{
\item{pdf_text}{PDF file.}
}
\value{
paper doi.
}
\description{
Extract DOI from PDF.
}
\examples{
\donttest{
pdf_doi_extract(pdf_text = pdf_text)
}
}
