% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extract.R
\name{pdf_github}
\alias{pdf_github}
\title{Detect and extract github link.}
\usage{
pdf_github(pdffile = NULL, git_word = NULL)
}
\arguments{
\item{pdffile}{PDF file name.}

\item{git_word}{github related word.}
}
\value{
Github links.
}
\description{
Detect and extract github link.
}
\examples{
\donttest{
setwd("C:/Users/huxin/Desktop/test")
fileinput <- list.files("C:/Users/huxin/Desktop/test")
fileinput
for (d in fileinput){
  try(pdf_github(pdffile = d))
}
}
}
