% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_extract.R
\name{pdf_reference_extract}
\alias{pdf_reference_extract}
\title{PDF reference extraction.}
\usage{
pdf_reference_extract(pdf_text = NULL, begin_word = NULL, end_word = NULL)
}
\arguments{
\item{pdf_text}{PDF file.}

\item{begin_word}{Words at the beginning of the reference section.}

\item{end_word}{Words at the end of the reference section.}
}
\value{
"pdf_reference_extract.txt".
}
\description{
PDF reference extraction.
}
\examples{
\donttest{
pdf_text <- pdf_text("Machine_learning_for_microbiologists.pdf")
begin_word <- "References|REFERENCES"
end_word <- "Acknowledgements"

pdf_reference_extract(pdf_text = pdf_text, begin_word = begin_word, end_word = end_word)
}

}
