% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coauthors_analysis.R
\name{plot_coauthors}
\alias{plot_coauthors}
\title{Draw a network diagram.}
\usage{
plot_coauthors(network, size_labels = 5)
}
\arguments{
\item{network}{The result obtained by running the build_author_network() function.}

\item{size_labels}{label size}
}
\value{
network diagram.
}
\description{
Draw a network diagram.
}
\examples{
\donttest{
plot_coauthors(g)
}
}
