% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_collection.R
\name{pubmed_extract}
\alias{pubmed_extract}
\title{Extract the search results from the PubMed database.}
\usage{
pubmed_extract(
  root =
    "https://pubmed.ncbi.nlm.nih.gov/?term=(Wang,\%20Xin[Author])\%20AND\%20(Zhang,\%20Yan[Author])&size=20&page=",
  page_number = 1,
  paper_numbers = 1
)
}
\arguments{
\item{root}{The public part of webpage links.}

\item{page_number}{Total pages of search results.}

\item{paper_numbers}{Total number of retrieved papers.}
}
\value{
A data frame including article title, abstract, etc.
}
\description{
Extract the search results from the PubMed database.
}
\examples{
\donttest{
pubmed_extract(page_number = 1, paper_numbers = 20)
}
}
