% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_download.R
\name{scihub_download}
\alias{scihub_download}
\title{Download PDF from sic-hub.}
\usage{
scihub_download(doi, file_dir)
}
\arguments{
\item{doi}{paper doi.}

\item{file_dir}{File save path.}
}
\value{
PDF.
}
\description{
Download PDF from sic-hub.
}
\examples{
\donttest{
doi <- final_table$DOI
for (i in doi) {
  print(i)
  try(scihub_download1(doi = i, file_dir = "C:/your/path/"), silent = TRUE)
}
}
}
