% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_link.R
\name{search_link}
\alias{search_link}
\title{Building a PubMed retrieval link.}
\usage{
search_link(term, size, page)
}
\arguments{
\item{term}{Search terms.}

\item{size}{The total number of papers contained on each page of the result.}

\item{page}{}
}
\value{
PubMed search link
}
\description{
Building a PubMed retrieval link.
}
\examples{
\donttest{
term <- "(m6A) AND (cancer)"
term <- "(Wang, Xin[Author]) AND (Zhang, Yan[Author])"
term <-  "((Wang, Xin[Author]) AND (Zhang, Yan[Author])) AND ((\"2023/01/02\"[Date - Publication] : \"2023/12/10\"[Date - Publication]))"
size <- "20"
page <- ''

search_link(term = term, size = size, page = page)
}
}
