% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_paper.R
\name{similar_literature}
\alias{similar_literature}
\title{similar literature recommend.}
\usage{
similar_literature(pubmedid = NULL)
}
\arguments{
\item{pubmedid}{The article PubMed ID.}
}
\value{
A data frame of similar literature information:similar_literature_table.
}
\description{
similar literature recommend.
}
\examples{
\donttest{
### Example
similar_literature(pubmedid = '32134010')
write.csv(similar_literature_table, file = 'similar_literature.csv',row.names = F)
}
}
