% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{year_plot}
\alias{year_plot}
\title{Counting the number of literature publications per year.}
\usage{
year_plot(year)
}
\arguments{
\item{year}{A data frame contain year of publication of the articles.}
}
\value{
line chart
}
\description{
Counting the number of literature publications per year.
}
\examples{
\dottest{
year_plot(final_table$Year)
}
}
